+++ src/Ryujinx.Tests/Audio/Renderer/Dsp/UpsamplerTests.cs
@@ -21,14 +21,14 @@ namespace Ryujinx.Tests.Audio.Renderer.Dsp
             for (int sample = 0; sample < inputBuffer.Length; sample++)
             {
                 // 440 hz sine wave with amplitude = 0.5f at input sample rate
-                inputBuffer[sample] = MathF.Sin((440 / inputSampleRate) * (float)sample * MathF.PI * 2f) * 0.5f;
+                inputBuffer[sample] = MathF.Sin((440 / inputSampleRate) * sample * MathF.PI * 2f) * 0.5f;
             }
 
             int inputIdx = 0;
             int outputIdx = 0;
             while (inputIdx + inputBlockSize < numInputSamples)
             {
-                int outputBufLength = (int)Math.Round((float)(inputIdx + inputBlockSize) * outputSampleRate / inputSampleRate) - outputIdx;
+                int outputBufLength = (int)Math.Round((inputIdx + inputBlockSize) * outputSampleRate / inputSampleRate) - outputIdx;
                 UpsamplerHelper.Upsample(
                     outputBuffer.AsSpan(outputIdx),
                     inputBuffer.AsSpan(inputIdx),
@@ -45,11 +45,11 @@ namespace Ryujinx.Tests.Audio.Renderer.Dsp
             for (int sample = 0; sample < numOutputSamples; sample++)
             {
                 // 440 hz sine wave with amplitude = 0.5f at output sample rate with an offset of 15
-                expectedOutput[sample] = MathF.Sin((440 / outputSampleRate) * (float)(sample - 15) * MathF.PI * 2f) * 0.5f;
+                expectedOutput[sample] = MathF.Sin((440 / outputSampleRate) * (sample - 15) * MathF.PI * 2f) * 0.5f;
                 sumDifference += Math.Abs(expectedOutput[sample] - outputBuffer[sample]);
             }
 
-            sumDifference /= (float)expectedOutput.Length;
+            sumDifference /= expectedOutput.Length;
             // Expect the output to be 98% similar to the expected resampled sine wave
             Assert.IsTrue(sumDifference < 0.02f);
         }
